% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_comp_dirichlet.R
\name{fit_comp_dirichlet}
\alias{fit_comp_dirichlet}
\title{Fit a Dirichlet regression model}
\usage{
fit_comp_dirichlet(data, formula)
}
\arguments{
\item{data}{A data frame containing the input data. Must include columns with names
starting with "alpha_sim" for the Dirichlet parameters.}

\item{formula}{A formula specifying the regression model to be fitted.}
}
\value{
A fitted Dirichlet regression model object.
}
\description{
This function fits a Dirichlet regression model using the \code{DirichletReg::DirichReg()}
function.
}
