% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics_comp.R
\name{mae_comp}
\alias{mae_comp}
\title{Mean Absolute Error (MAE) for compositional data}
\usage{
mae_comp(truth, estimate, summarize = FALSE)
}
\arguments{
\item{truth}{A matrix or data frame of true values.}

\item{estimate}{A matrix or data frame of estimated values.}

\item{summarize}{Logical. If \code{TRUE}, returns a single summarized MAE value.
If \code{FALSE}, returns MAE values for each component.}
}
\value{
A tibble containing the MAE values. If \code{summarize = TRUE}, the tibble contains
one row with the mean MAE. If \code{summarize = FALSE}, the tibble contains detailed
MAE values for each component.
}
\description{
This function computes the MAE between true and estimated values for compositional data.
}
