% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics_comp.R
\name{mtr}
\alias{mtr}
\title{Median true result when estimated value is greater than or equal to some threshold value (implemented as a custom performance metric in the \href{https://yardstick.tidymodels.org/}{yardstick} package)}
\usage{
mtr(data, ...)
}
\arguments{
\item{data}{A \code{data.frame} containing the columns specified by the truth and estimate arguments.}

\item{...}{Not currently used.}

\item{truth}{The column identifier for the true results (that is \code{numeric}). This should be an unquoted column name although this argument is passed by expression and supports quasiquotation (you can unquote column names). For \verb{_vec()}⁠ functions, a \code{numeric} vector.}

\item{estimate}{The column identifier for the predicted results (that is also \code{numeric}). As with \code{truth} this can be specified different ways but the primary method is to use an unquoted variable name. For \verb{_vec()}⁠ functions, a numeric \code{vector}.}

\item{thresh}{Threshold value}

\item{na_rm}{A \code{logical} value indicating whether \code{NA} values should be stripped before the computation proceeds.}

\item{case_weights}{The optional column identifier for case weights. This should be an unquoted column name that evaluates to a numeric column in \code{data}. For \verb{_vec()} functions, a \code{numeric} vector.}
}
\value{
description
}
\description{
Median true result when estimated value is greater than or equal to some threshold value (implemented as a custom performance metric in the \href{https://yardstick.tidymodels.org/}{yardstick} package)
}
