% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_comp_data.R
\name{plot_comp_data}
\alias{plot_comp_data}
\title{Plot simulated compositional data}
\usage{
plot_comp_data(sim)
}
\arguments{
\item{sim}{description}
}
\value{
description
}
\description{
Plot simulated compositional data
}
