% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_compositional_data.R
\name{predict_compositional_data}
\alias{predict_compositional_data}
\title{Predict substrate composition probabilities}
\usage{
predict_compositional_data(object, new_data)
}
\arguments{
\item{object}{A fitted Dirichlet regression model object of class \code{DirichletRegModel} (from \code{DirichletReg}).}

\item{new_data}{A data frame or a \code{PackedSpatRaster}/\code{SpatRaster} object containing the input data.}
}
\value{
A modified version of the input \code{data} object, including predicted
substrate composition probabilities and model-estimated alpha values.
If the input is a \code{SpatRaster}, the output is a wrapped raster object.
}
\description{
This function predicts substrate composition probabilities based on a fitted
Dirichlet regression model.
}
