% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_comp_data.R
\name{simulate_comp_data}
\alias{simulate_comp_data}
\title{Simulate compositional data from a Dirichlet likelihood}
\usage{
simulate_comp_data(
  x,
  d,
  layers = NULL,
  poly_degree = NULL,
  n_cov_sim = NULL,
  as_raster = FALSE,
  seed = NULL
)
}
\arguments{
\item{x}{description}

\item{d}{description}

\item{layers}{description}

\item{poly_degree}{description}

\item{n_cov_sim}{description}

\item{as_raster}{description}

\item{seed}{description}
}
\value{
description
}
\description{
Simulate compositional data from a Dirichlet likelihood
}
