% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_comp_dirichlet.R
\name{fit_comp_dirichlet}
\alias{fit_comp_dirichlet}
\title{Fit dirichlet regression model (via dirinla::dirinlareg() function)}
\usage{
fit_comp_dirichlet(
  data,
  formula,
  use_dirinla = FALSE,
  tol0 = NULL,
  verbose = FALSE
)
}
\arguments{
\item{data}{description}

\item{formula}{description}

\item{use_dirinla}{description}

\item{tol0}{description}

\item{verbose}{description}
}
\value{
description
}
\description{
Fit dirichlet regression model (via dirinla::dirinlareg() function)
}
