% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics_comp.R
\name{metrics_comp}
\alias{metrics_comp}
\title{Performance metrics for compositional data}
\usage{
metrics_comp(truth, estimate, metric = "all", summarize = FALSE)
}
\arguments{
\item{truth}{description}

\item{estimate}{description}

\item{metric}{description}

\item{summarize}{description}
}
\value{
description
}
\description{
Performance metrics for compositional data
}
