% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_compositional_data.R
\name{plot_compositional_data}
\alias{plot_compositional_data}
\title{Plot simulated compositional rasters}
\usage{
plot_compositional_data(sim)
}
\arguments{
\item{sim}{A list containing the simulated data. The \code{sim$data} element must
be a \code{SpatRaster} or \code{PackedSpatRaster} object. The \code{sim$d} element
specifies the number of compositional components to plot.}
}
\value{
Invisibly returns \code{sim}. Produces a multi-panel plot as side effect.
}
\description{
This function generates a plot of simulated compositional data
stored in a \code{SpatRaster} object. It extracts the relevant layers from the
raster object and visualizes them as a series of plots.
}
\examples{
\donttest{
library(terra)
r <- rast(system.file("ex/elev.tif", package = "terra"))
sim <- simulate_compositional_data(r, d = 3, as_raster = TRUE)
plot_compositional_data(sim)
}

}
\seealso{
\code{\link[=simulate_compositional_data]{simulate_compositional_data()}}
}
