% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_comp_dirichlet.R
\name{predict_comp_dirichlet}
\alias{predict_comp_dirichlet}
\title{Predict substrate composition values based on dirichlet regression model}
\usage{
predict_comp_dirichlet(data, model)
}
\arguments{
\item{data}{description}

\item{model}{description}
}
\value{
description
}
\description{
Predict substrate composition values based on dirichlet regression model
}
