% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_comp_dirichlet.R
\name{predict_comp_dirichlet}
\alias{predict_comp_dirichlet}
\title{Predict substrate composition probabilities}
\usage{
predict_comp_dirichlet(data, model)
}
\arguments{
\item{data}{A data frame or a \code{PackedSpatRaster}/\code{SpatRaster} object containing the input data.}

\item{model}{A fitted Dirichlet regression model object, either of class
\code{dirinlaregmodel} (from \code{dirinla}) or \code{DirichletRegModel} (from \code{DirichletReg}).}
}
\value{
A modified version of the input \code{data} object, including predicted
substrate composition probabilities and model-estimated alpha values.
If the input is a \code{SpatRaster}, the output is a wrapped raster object.
}
\description{
This function predicts substrate composition probabilities based on a fitted
Dirichlet regression model.
}
