% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_simulation.R
\name{run_simulation}
\alias{run_simulation}
\title{Run simulation}
\usage{
run_simulation(
  sim_data,
  sites = NULL,
  formula,
  n,
  method,
  strata_var = NULL,
  num_sim,
  use_dirinla = TRUE,
  tol0 = NULL,
  verbose = FALSE
)
}
\arguments{
\item{sim_data}{description}

\item{sites}{description}

\item{formula}{description}

\item{n}{description}

\item{method}{description}

\item{strata_var}{description}

\item{num_sim}{description}

\item{use_dirinla}{description}

\item{tol0}{description}

\item{verbose}{description}
}
\value{
description
}
\description{
Run simulation
}
