% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_comp_dirichlet.R
\name{sample_fit_comp_dirichlet}
\alias{sample_fit_comp_dirichlet}
\title{Sample the data and fit dirichlet regression model (via dirinla:: function)}
\usage{
sample_fit_comp_dirichlet(
  data,
  n,
  method,
  bias_var = NULL,
  bias_thresh = NULL,
  clh_var = NULL,
  clh_iter = NULL,
  strata_var = NULL,
  formula
)
}
\arguments{
\item{data}{description}

\item{n}{description}

\item{method}{description}

\item{bias_var}{description}

\item{bias_thresh}{description}

\item{clh_var}{description}

\item{clh_iter}{description}

\item{strata_var}{description}

\item{formula}{description}
}
\value{
description
}
\description{
Sample the data and fit dirichlet regression model (via dirinla:: function)
}
